third_party/blink/web_tests/external/wpt/infrastructure/server/subdomain-flag.www.sub.window-expected.txt

This is a testharness.js-based test.
FAIL Use of .www. file name flag implies www subdomain assert_equals: expected "www.web-platform.test" but got "web-platform.test"
Harness: the test ran to completion.
